Rename project to best practices.
This better reflects the purpose of the repository.
