+++ .github/workflows/release.yml
@@ -4,8 +4,8 @@ on: push
 
 env:
   # Change these for your project's URLs
-  PYPI_URL: https://pypi.org/p/django-commons-playground
-  PYPI_TEST_URL: https://test.pypi.org/p/django-commons-playground
+  PYPI_URL: https://pypi.org/p/django-commons-best-practices
+  PYPI_TEST_URL: https://test.pypi.org/p/django-commons-best-practices
 
 jobs:
