+++ CODE_OF_CONDUCT.md
@@ -1,3 +1,3 @@
-# Django Commons Playground Code of Conduct
+# Code of Conduct
 
-The django-commons-playground project utilizes the [Django Commons Code of Conduct](https://github.com/django-commons/membership/blob/main/CODE_OF_CONDUCT.md).
+The best-practices project utilizes the [Django Commons Code of Conduct](https://github.com/django-commons/membership/blob/main/CODE_OF_CONDUCT.md).
