+++ README.md
@@ -2,5 +2,5 @@
 
 This is an example repository for Django Commons. Some relevant areas to review are:
 
-- GitHub action ([release.yml](https://github.com/django-commons/django-commons-playground/blob/main/.github/workflows/release.yml)) for releasing to PyPI
+- GitHub action ([release.yml](https://github.com/django-commons/best-practices/blob/main/.github/workflows/release.yml)) for releasing to PyPI
 - [Documentation](docs/README.md)
