+++ docs/README.md
@@ -1,6 +1,6 @@
 # Documentation
 
-The documentation for Django Commons Playground serves as an example for what
+The documentation for Django Commons Best Practices serves as an example for what
 a repository should contain. The structure doesn't have to be exactly the same
 as this, but it's a starting point. Any docs is better than no docs. Maintained
 and organized docs are the best docs.
