+++ docs/installation.md
@@ -1,16 +1,16 @@
 # Installation
 
-This section explains how to install `django-commons-playground` to use within
+This section explains how to install `django-commons-best-practices` to use within
 your project. It assumes you have a working installation of Python 3.
 
 1. Create venv/virtualenv (`python -m venv venv`)
 2. Activate venv/virtualenv (`source venv/bin/activate`)
-3. Install `django-commons-playground` (`pip install django-commons-playground`)
+3. Install `django-commons-best-practices` (`pip install django-commons-best-practices`)
 
 All together now:
 
 ```shell
 python -m venv venv
 source venv/bin/activate
-pip install django-commons-playground
+pip install django-commons-best-practices
 ```
