+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools>=61.0"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "django_commons_playground"
+name = "django-commons-best-practices"
 version = "1.1.0"
 authors = [
   { name="Tim Schilling", email="schillingt@better-simple.com" },
@@ -22,5 +22,5 @@ keywords = [
 ]
 
 [project.urls]
-Homepage = "https://github.com/django-commons/django-commons-playground"
-Issues = "https://github.com/django-commons/django-commons-playground/issues"
+Homepage = "https://github.com/django-commons/best-practices"
+Issues = "https://github.com/django-commons/best-practices/issues"
